requirements.txt
# Environment and Configuration
python-dotenv
pyyaml

# LLM and AI Framework
langchain
langchain-community
langchain-openai
langchain-google-genai
google-generativeai
openai

# Database
neo4j

# Local LLM
ollama

# HTTP and API
requests

# Jupyter and Development
ipykernel
notebook

# Data Processing and Analysis
numpy
pandas
matplotlib

# Testing
pytest